WC-621: client.ops.his: Handle rng=slice object case.
MeterMaster passes in a `slice` object with two `datetime.datetime`
objects when it wishes to do a read operation.  It seems this construct
causes some grief with `pyhaystack` and `hszinc` which shouldn't be an
issue.

This fixes the issue by converting the slice object into a
comma-separated string: standard Project Haystack format.