util.scram: Add function to convert integer to bytes.
It seems with `_xor`, it returns hex *only* because we call `unhexlify`
on it to get a bytestring in big-endian format for use with the `base64`
library.

My first instinct is to use `struct`, but I don't know how bug the
integer is, so best bet, let's not assume, we'll pack it ourselves.
This takes an unsigned integer of arbitrary size (8-bits, 128-bits, who
cares?) and outputs the byte string ready for consumption by `base64`.